Se suben actividades

+++ Java/Autonomo/src/main/java/com/example/Autonomo/Service/ProductoService.java
@@ -4,9 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.example.Autonomo.Entity.Producto;
+import com.example.Autonomo.Repository.OrdenDeCompraRepository;
 import com.example.Autonomo.Repository.ProductoRepository;
 
 @Service
@@ -15,6 +18,9 @@ public class ProductoService {
     @Autowired
     private ProductoRepository productoRepository;
 
+    @Autowired
+    private OrdenDeCompraRepository ordenDeCompraRepository;
+
     public List<Producto> getAllProductos() {
         return productoRepository.findAll();
     }
@@ -22,6 +28,7 @@ public class ProductoService {
     public Optional<Producto> getProductoById(Long Id) {
         return productoRepository.findById(Id);
     }
+
     public Producto createProducto(Producto producto) {
         return productoRepository.save(producto);
     }
@@ -34,7 +41,11 @@ public class ProductoService {
         return null;
     }
 
+    @Transactional
     public void deleteProducto(Long Id) {
+        if (ordenDeCompraRepository.existsByProductoIdProducto(Id)) {
+            throw new DataIntegrityViolationException("El producto no puede ser eliminado porque está asociado a una orden de compra.");
+        }
         productoRepository.deleteById(Id);
     }
 }
