+++ Java/Autonomo/src/main/java/com/example/Autonomo/Service/ProveedorService.java
@@ -4,9 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.example.Autonomo.Entity.Proveedor;
+import com.example.Autonomo.Repository.ProductoRepository;
 import com.example.Autonomo.Repository.ProveedorRepository;
 
 @Service
@@ -15,22 +18,21 @@ public class ProveedorService {
     @Autowired
     private ProveedorRepository proveedorRepository;
 
-    // Obtener todos los proveedores
+    @Autowired
+    private ProductoRepository productoRepository;
+
     public List<Proveedor> getAllProveedores() {
         return proveedorRepository.findAll();
     }
 
-    // Obtener un proveedor por su ID
     public Optional<Proveedor> getProveedorById(Long Id) {
         return proveedorRepository.findById(Id);
     }
 
-    // Crear un nuevo proveedor
     public Proveedor createProveedor(Proveedor proveedor) {
         return proveedorRepository.save(proveedor);
     }
 
-    // Actualizar un proveedor
     public Proveedor updateProveedor(Long Id, Proveedor proveedor) {
         if (proveedorRepository.existsById(Id)) {
             proveedor.setIdProveedor(Id);
@@ -39,8 +41,11 @@ public class ProveedorService {
         return null;
     }
 
-    // Eliminar un proveedor
+    @Transactional
     public void deleteProveedor(Long Id) {
+        if (productoRepository.existsByProveedorIdProveedor(Id)) {
+            throw new DataIntegrityViolationException("El proveedor no puede ser eliminado porque está asociado a uno o más productos.");
+        }
         proveedorRepository.deleteById(Id);
     }
 }
